#1342: split forwarding_build.rs by entity kind
The 1162-LOC userspace-dp/src/afxdp/forwarding_build.rs hosted two
god-function violations: build_forwarding_state at 469 LOC and
build_cos_state at 312 LOC, both well above engineering-style.md's
>100 LOC refactor cue and the >200 LOC tier-1 refresh-audit
threshold.

Decompose into a directory module:

  userspace-dp/src/afxdp/forwarding_build/
    mod.rs        - orchestrator + screen-profile + syn-cookie +
                    late-stage NAT/DNAT local-delivery (345L)
    zones.rs      - populate_zones (46L)
    tunnels.rs    - populate_tunnel_endpoints (53L)
    interfaces.rs - populate_interfaces + populate_egress +
                    pick_interface_v4/v6 + IfaceIndex context
                    (215L)
    fib.rs        - sort_connected + populate_routes + sort_routes
                    + populate_neighbors + populate_fabrics +
                    route-target / next-hop / ifindex resolvers
                    (289L)
    cos.rs        - build_cos_classifier_tables +
                    build_cos_iface_config + build_cos_state
                    orchestrator + CoS table builders (470L)
    tests.rs      - moved from forwarding_build_tests.rs

build_cos_state is further decomposed into three helpers per
#1342 r1 review (Codex r1-MINOR-5, AGY r1-#6): classifier-table
construction, per-interface config build (which owns the #1183
useful_cos_state gate), and a slim orchestrator under 30 LOC.

Pure code motion. Zero behaviour change.

Visibility:
- pub(in crate::afxdp) for helpers consumed by other afxdp
  siblings via use self::forwarding_build::*;: pick_interface_v4,
  pick_interface_v6, resolve_route_target_v4/v6,
  parse_route_next_hop{,_v6}, resolve_ifindex,
  infer_connected_route_target_v4/v6.
- pub(super) for orchestrator-local: build_forwarding_state*,
  build_screen_profiles, build_cos_state, default_cos_burst_bytes.
- Plain (private) use cos::{...} in mod.rs avoids the E0364 that
  pub(super) use of pub(super) item would trigger (see
  tx/mod.rs:38 precedent).

Side-effect ordering preserved exactly:
- zones -> tunnels -> interfaces(addresses) -> interfaces(egress)
  -> sort_connected -> routes -> sort_routes -> neighbors ->
  fabrics -> policy -> flow knobs/NAT/screens/MSS/filter -> cos
  -> tx_selection_v4/v6 -> flow_export -> mirror -> static-NAT
  local-delivery -> DNAT local-delivery -> debug-log block ->
  install_kernel_rst_suppression.
- static-NAT and DNAT local-delivery loops stay in mod.rs at
  their late position (AGY r1 #2): moving them earlier would
  execute before state.static_nat / state.dnat_table are
  populated, silently emptying NAT local-delivery sets.
- useful_cos_state gate retained bit-for-bit inside
  build_cos_iface_config, ordered after queue-resolution and
  before the synthetic best-effort fallback (AGY r1 #5).

IfaceIndex context struct (name_to_ifindex, linux_to_ifindex,
mac_by_ifindex) lifts the three locals previously rebuilt across
the interfaces / egress / routes / fabrics walks into a typed
shared context, holding the orchestrator under the 8-param cap.
Owned struct, no lifetimes, no borrow conflicts.

ClassifierTables<'a> in cos.rs mirrors the borrow-only scheduler
/ scheduler-map maps that the original code built inline. No new
allocations; same FastMap shapes.

Plan + adversarial review:
- Plan doc: docs/pr/1342-forwarding-build-split/plan.md v2.
- Codex round-1: PLAN-NEEDS-MAJOR (3 majors, 2 minors).
- Codex round-2: PLAN-NEEDS-MINOR (default_cos_burst_bytes
  test reach) - addressed in implementation by `use cos::{...,
  default_cos_burst_bytes};` in mod.rs.
- AGY round-1: PLAN-NEEDS-MINOR (5 action items).
- AGY round-2: PLAN-READY.

Test plan:
- cargo build --release clean.
- cargo test --release: 1416 pass; 1 pre-existing unrelated
  flake (afxdp::wg::engine::reconcile_peers_snapshot_is_atomic)
  verified 5x clean on retry.
- 5/5 flake check on
  build_forwarding_state_prefers_logical_unit_for_ingress_lookup
  clean.
- go test ./...: clean.
- Wave-2 refactor batch: per-PR smoke skipped per the AWAITING-
  BATCH-MERGE convention.

Closes #1342
