+++ docs/pr/1342-forwarding-build-split/reviewer-ids.md
@@ -5,8 +5,13 @@ instead of re-dispatching.
 
 ## Plan review round 1
 
-- Codex: <pending>
-- AGY: <pending>
+- Codex: task-mpmyyj1v-k8y4kb — PLAN-NEEDS-MAJOR (3 majors, 2 minors)
+- AGY: adversarial-review-mpmyyibe-upfkb0 — PLAN-NEEDS-MINOR (5 items)
+
+## Plan review round 2
+
+- Codex: task-mpmzvqal-0o1rq0
+- AGY: adversarial-review-mpmzvovv-rfrol2
 
 ## Code review round 1
