+++ userspace-dp/src/afxdp/forwarding_build/fib.rs
@@ -1,0 +1,289 @@
+//! FIB / neighbor / fabric population for `build_forwarding_state`.
+//!
+//! Owns:
+//!
+//! - [`sort_connected`] — sort `state.connected_v[46]` by prefix length.
+//! - [`populate_routes`] — walk `snapshot.routes`, push into
+//!   `state.routes_v[46]` keyed by canonical table name.
+//! - [`sort_routes`] — sort each route table by prefix length.
+//! - [`populate_neighbors`] — copy usable neighbors into
+//!   `state.neighbors`.
+//! - [`populate_fabrics`] — append `FabricLink` entries to
+//!   `state.fabrics`, resolving local + peer MACs through
+//!   [`IfaceIndex`] and the populated `state.neighbors` map.
+//!
+//! Also hosts the route-target resolution helpers
+//! ([`resolve_route_target_v4`] etc.) re-exported by
+//! `forwarding_build/mod.rs` as `pub(in crate::afxdp)` for the
+//! other afxdp siblings that consume them via `use
+//! self::forwarding_build::*;` in `afxdp/mod.rs`.
+
+use super::super::*;
+use super::interfaces::IfaceIndex;
+use crate::RouteSnapshot;
+use ipnet::{Ipv4Net, Ipv6Net};
+use std::collections::BTreeMap;
+use std::net::{IpAddr, Ipv4Addr, Ipv6Addr};
+
+pub(super) fn sort_connected(state: &mut ForwardingState) {
+    state
+        .connected_v4
+        .sort_by(|a, b| b.prefix.prefix_len().cmp(&a.prefix.prefix_len()));
+    state
+        .connected_v6
+        .sort_by(|a, b| b.prefix.prefix_len().cmp(&a.prefix.prefix_len()));
+}
+
+pub(super) fn populate_routes(
+    snapshot: &ConfigSnapshot,
+    state: &mut ForwardingState,
+    iface_ctx: &IfaceIndex,
+) {
+    for route in &snapshot.routes {
+        if let Ok(prefix) = route.destination.parse::<Ipv4Net>() {
+            let (next_hop, ifindex, tunnel_endpoint_id) = resolve_route_target_v4(
+                route,
+                &iface_ctx.name_to_ifindex,
+                &iface_ctx.linux_to_ifindex,
+                state,
+            );
+            let table = canonical_route_table(&route.table, false);
+            state
+                .routes_v4
+                .entry(table)
+                .or_default()
+                .push(RouteEntryV4 {
+                    prefix: PrefixV4::from_net(prefix),
+                    ifindex,
+                    tunnel_endpoint_id,
+                    next_hop,
+                    discard: route.discard,
+                    next_table: route.next_table.clone(),
+                });
+            continue;
+        }
+        if let Ok(prefix) = route.destination.parse::<Ipv6Net>() {
+            let (next_hop, ifindex, tunnel_endpoint_id) = resolve_route_target_v6(
+                route,
+                &iface_ctx.name_to_ifindex,
+                &iface_ctx.linux_to_ifindex,
+                state,
+            );
+            let table = canonical_route_table(&route.table, true);
+            state
+                .routes_v6
+                .entry(table)
+                .or_default()
+                .push(RouteEntryV6 {
+                    prefix: PrefixV6::from_net(prefix),
+                    ifindex,
+                    tunnel_endpoint_id,
+                    next_hop,
+                    discard: route.discard,
+                    next_table: route.next_table.clone(),
+                });
+        }
+    }
+}
+
+pub(super) fn sort_routes(state: &mut ForwardingState) {
+    for routes in state.routes_v4.values_mut() {
+        routes.sort_by(|a, b| b.prefix.prefix_len().cmp(&a.prefix.prefix_len()));
+    }
+    for routes in state.routes_v6.values_mut() {
+        routes.sort_by(|a, b| b.prefix.prefix_len().cmp(&a.prefix.prefix_len()));
+    }
+}
+
+pub(super) fn populate_neighbors(snapshot: &ConfigSnapshot, state: &mut ForwardingState) {
+    for neigh in &snapshot.neighbors {
+        if neigh.ifindex <= 0 || !neighbor_state_usable(&neigh.state) {
+            continue;
+        }
+        let Ok(ip) = neigh.ip.parse::<IpAddr>() else {
+            continue;
+        };
+        let Some(mac) = parse_mac(&neigh.mac) else {
+            continue;
+        };
+        state
+            .neighbors
+            .insert((neigh.ifindex, ip), NeighborEntry { mac });
+    }
+}
+
+pub(super) fn populate_fabrics(
+    snapshot: &ConfigSnapshot,
+    state: &mut ForwardingState,
+    iface_ctx: &IfaceIndex,
+) {
+    for fabric in &snapshot.fabrics {
+        if fabric.parent_ifindex <= 0 {
+            continue;
+        }
+        let Ok(peer_addr) = fabric.peer_address.parse::<IpAddr>() else {
+            continue;
+        };
+        let local_mac = parse_mac(&fabric.local_mac)
+            .or_else(|| iface_ctx.mac_by_ifindex.get(&fabric.parent_ifindex).copied());
+        let Some(local_mac) = local_mac else {
+            continue;
+        };
+        let peer_mac = parse_mac(&fabric.peer_mac).or_else(|| {
+            state
+                .neighbors
+                .get(&(fabric.overlay_ifindex, peer_addr))
+                .or_else(|| state.neighbors.get(&(fabric.parent_ifindex, peer_addr)))
+                .map(|entry| entry.mac)
+        });
+        let Some(peer_mac) = peer_mac else {
+            continue;
+        };
+        state.fabrics.push(FabricLink {
+            parent_ifindex: fabric.parent_ifindex,
+            overlay_ifindex: fabric.overlay_ifindex,
+            peer_addr,
+            peer_mac,
+            local_mac,
+        });
+    }
+}
+
+pub(in crate::afxdp) fn resolve_route_target_v4(
+    route: &RouteSnapshot,
+    names: &BTreeMap<String, i32>,
+    linux_names: &BTreeMap<String, i32>,
+    state: &ForwardingState,
+) -> (Option<Ipv4Addr>, i32, u16) {
+    if route.discard || !route.next_table.is_empty() {
+        return (None, 0, 0);
+    }
+    let Some((next_hop, interface)) = route
+        .next_hops
+        .first()
+        .map(|nh| parse_route_next_hop(nh.as_str()))
+    else {
+        return (None, 0, 0);
+    };
+    let target = interface
+        .as_deref()
+        .and_then(|name| resolve_ifindex(name, names, linux_names))
+        .map(|ifindex| {
+            (
+                ifindex,
+                state
+                    .tunnel_endpoint_by_ifindex
+                    .get(&ifindex)
+                    .copied()
+                    .unwrap_or(0),
+            )
+        })
+        .or_else(|| next_hop.and_then(|ip| infer_connected_route_target_v4(state, ip)));
+    let (ifindex, tunnel_endpoint_id) = target.unwrap_or((0, 0));
+    (next_hop, ifindex, tunnel_endpoint_id)
+}
+
+pub(in crate::afxdp) fn resolve_route_target_v6(
+    route: &RouteSnapshot,
+    names: &BTreeMap<String, i32>,
+    linux_names: &BTreeMap<String, i32>,
+    state: &ForwardingState,
+) -> (Option<Ipv6Addr>, i32, u16) {
+    if route.discard || !route.next_table.is_empty() {
+        return (None, 0, 0);
+    }
+    let Some((next_hop, interface)) = route
+        .next_hops
+        .first()
+        .map(|nh| parse_route_next_hop_v6(nh.as_str()))
+    else {
+        return (None, 0, 0);
+    };
+    let target = interface
+        .as_deref()
+        .and_then(|name| resolve_ifindex(name, names, linux_names))
+        .map(|ifindex| {
+            (
+                ifindex,
+                state
+                    .tunnel_endpoint_by_ifindex
+                    .get(&ifindex)
+                    .copied()
+                    .unwrap_or(0),
+            )
+        })
+        .or_else(|| next_hop.and_then(|ip| infer_connected_route_target_v6(state, ip)));
+    let (ifindex, tunnel_endpoint_id) = target.unwrap_or((0, 0));
+    (next_hop, ifindex, tunnel_endpoint_id)
+}
+
+pub(in crate::afxdp) fn parse_route_next_hop(spec: &str) -> (Option<Ipv4Addr>, Option<String>) {
+    let (ip_part, if_part) = if let Some((lhs, rhs)) = spec.split_once('@') {
+        (lhs, rhs)
+    } else {
+        (spec, "")
+    };
+    let ip = if ip_part.is_empty() {
+        None
+    } else {
+        ip_part.parse::<Ipv4Addr>().ok()
+    };
+    let iface = if if_part.is_empty() {
+        None
+    } else {
+        Some(if_part.to_string())
+    };
+    (ip, iface)
+}
+
+pub(in crate::afxdp) fn parse_route_next_hop_v6(spec: &str) -> (Option<Ipv6Addr>, Option<String>) {
+    let (ip_part, if_part) = if let Some((lhs, rhs)) = spec.split_once('@') {
+        (lhs, rhs)
+    } else {
+        (spec, "")
+    };
+    let ip = if ip_part.is_empty() {
+        None
+    } else {
+        ip_part.parse::<Ipv6Addr>().ok()
+    };
+    let iface = if if_part.is_empty() {
+        None
+    } else {
+        Some(if_part.to_string())
+    };
+    (ip, iface)
+}
+
+pub(in crate::afxdp) fn resolve_ifindex(
+    name: &str,
+    names: &BTreeMap<String, i32>,
+    linux_names: &BTreeMap<String, i32>,
+) -> Option<i32> {
+    names
+        .get(name)
+        .copied()
+        .or_else(|| linux_names.get(name).copied())
+}
+
+pub(in crate::afxdp) fn infer_connected_route_target_v4(
+    state: &ForwardingState,
+    ip: Ipv4Addr,
+) -> Option<(i32, u16)> {
+    state
+        .connected_v4
+        .iter()
+        .find(|entry| entry.prefix.contains(ip))
+        .map(|entry| (entry.ifindex, entry.tunnel_endpoint_id))
+}
+
+pub(in crate::afxdp) fn infer_connected_route_target_v6(
+    state: &ForwardingState,
+    ip: Ipv6Addr,
+) -> Option<(i32, u16)> {
+    state
+        .connected_v6
+        .iter()
+        .find(|entry| entry.prefix.contains(ip))
+        .map(|entry| (entry.ifindex, entry.tunnel_endpoint_id))
+}
